+++ VueForm/ConnectedCheckbox.js
@@ -25,6 +25,7 @@ export default {
     indeterminate: String,
 
     validators: Array,
+    asyncValidators: Array,
 
     handleFocus: {
       type: Function,
@@ -49,7 +50,7 @@ export default {
   data() {
     const $registerFormComponent = resolveRegisterFormComponent(this)
 
-    return $registerFormComponent(this.name, this.value, this.validators)
+    return $registerFormComponent(this.name, this.value, this.validators, this.asyncValidators)
   },
 
   destroyed() {