+++ VueForm/ConnectedDatePicker.js
@@ -73,6 +73,7 @@ export default {
     clearIcon: String,
 
     validators: Array,
+    asyncValidators: Array,
 
     handleFocus: {
       type: Function,
@@ -98,7 +99,7 @@ export default {
   data() {
     const $registerFormComponent = resolveRegisterFormComponent(this)
 
-    return $registerFormComponent(this.name, this.value, this.validators)
+    return $registerFormComponent(this.name, this.value, this.validators, this.asyncValidators)
   },
 
   destroyed() {