+++ VueForm/ConnectedRadio.js
@@ -17,6 +17,7 @@ export default {
     size: String,
 
     validators: Array,
+    asyncValidators: Array,
 
     handleFocus: {
       type: Function,
@@ -41,7 +42,7 @@ export default {
   data() {
     const $registerFormComponent = resolveRegisterFormComponent(this)
 
-    return $registerFormComponent(this.name, this.value, this.validators)
+    return $registerFormComponent(this.name, this.value, this.validators, this.asyncValidators)
   },
 
   destroyed() {